# services/api/openapi.json
{
  "components": {
    "schemas": {
...
      "CaseStatus": {
        "description": "`cases.status`.",
        "enum": [
          "running",
          "decided",
          "no_ai_decision",
          "closed"
        ],
        "title": "CaseStatus",
        "type": "string"
      },
...
      "CaseVerdict": {
        "enum": [
          "tp",
          "fp",
          "suspicious"
        ],
        "title": "CaseVerdict",
        "type": "string"
      },
...
      "FeedbackAnswer": {
        "additionalProperties": false,
        "description": "`POST /cases/{case_id}/feedback`: the stored feedback, with who wrote it.",
        "properties": {
          "case_id": {
            "title": "Case Id",
            "type": "string"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "reason": {
            "$ref": "#/components/schemas/FeedbackReason"
          },
          "user_subject": {
            "title": "User Subject",
            "type": "string"
          },
          "verdict": {
            "$ref": "#/components/schemas/CaseVerdict"
          }
        },
        "required": [
          "case_id",
          "user_subject",
          "verdict",
          "reason",
          "created_at"
        ],
        "title": "FeedbackAnswer",
        "type": "object"
      },
      "FeedbackReason": {
        "enum": [
          "correct",
          "was_tp_not_fp",
          "was_fp_not_tp",
          "missing_context",
          "wrong_urgent_events",
          "other"
        ],
        "title": "FeedbackReason",
        "type": "string"
      },
      "FieldError": {
        "additionalProperties": false,
        "properties": {
          "field": {
            "title": "Field",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "field",
          "message"
        ],
        "title": "FieldError",
        "type": "object"
      },
...
      "Level": {
        "description": "Severity level. Values compare as strings, not by severity.",
        "enum": [
          "low",
          "medium",
          "high",
          "critical"
        ],
        "title": "Level",
        "type": "string"
      },
...
      "Page_QAItemSummary_": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/QAItemSummary"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items"
        ],
        "title": "Page[QAItemSummary]",
        "type": "object"
      },
...
      "Problem": {
        "description": "The shape of an RFC 9457 answer, for the UI's own error handling.\n\nNot a response model of any route: `ais0c_api.openapi` makes it every operation's `default`\nresponse in the schema. `errors`, `domains` and `list_names` appear only on the problems that\ncarry them.",
        "properties": {
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Detail"
          },
          "domains": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Domains"
          },
          "errors": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/FieldError"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Errors"
          },
          "list_names": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "List Names"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "default": "about:blank",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "title",
          "status"
        ],
        "title": "Problem",
        "type": "object"
      },
      "QAItemSummary": {
        "additionalProperties": false,
        "description": "One row of the QA queue.",
        "properties": {
          "case_id": {
            "title": "Case Id",
            "type": "string"
          },
          "case_notify_level": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Level"
              },
              {
                "type": "null"
              }
            ]
          },
          "case_status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CaseStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "case_summary_tr": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Case Summary Tr"
          },
          "case_verdict": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CaseVerdict"
              },
              {
                "type": "null"
              }
            ]
          },
          "evaluation_no": {
            "title": "Evaluation No",
            "type": "integer"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "reason": {
            "$ref": "#/components/schemas/QAReason"
          },
          "resolved_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolved At"
          },
          "resolved_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolved By"
          },
          "status": {
            "$ref": "#/components/schemas/QAStatus"
          }
        },
        "required": [
          "id",
          "case_id",
          "evaluation_no",
          "reason",
          "status",
          "resolved_by",
          "resolved_at"
        ],
        "title": "QAItemSummary",
        "type": "object"
      },
      "QAReason": {
        "enum": [
          "random_sample",
          "verifier_conflict",
          "injection_suspected",
          "low_confidence",
          "fp_with_data_gap"
        ],
        "title": "QAReason",
        "type": "string"
      },
      "QAResolveAnswer": {
        "additionalProperties": false,
        "description": "The resolved item and the feedback it wrote for the case.",
        "properties": {
          "feedback": {
            "$ref": "#/components/schemas/FeedbackAnswer"
          },
          "item": {
            "$ref": "#/components/schemas/QAItemSummary"
          }
        },
        "required": [
          "item",
          "feedback"
        ],
        "title": "QAResolveAnswer",
        "type": "object"
      },
      "QAResolveRequest": {
        "additionalProperties": false,
        "description": "`POST /qa/{id}/resolve`: the operator's verdict, reason and optional comment.",
        "properties": {
          "comment": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          },
          "reason": {
            "$ref": "#/components/schemas/FeedbackReason"
          },
          "verdict": {
            "$ref": "#/components/schemas/CaseVerdict"
          }
        },
        "required": [
          "verdict",
          "reason"
        ],
        "title": "QAResolveRequest",
        "type": "object"
      },
      "QAStatus": {
        "description": "`qa_items.status`.",
        "enum": [
          "open",
          "resolved"
        ],
        "title": "QAStatus",
        "type": "string"
      },
...
    }
  },
  "info": {
    "description": "The analyst UI's API. The UI talks only to this service.\n\nErrors are RFC 9457 problems whose `title` is a machine-readable code (`catalog.rule_not_found`,\n`qa.already_resolved`); the Turkish message the user reads is built by the UI from that code.\nTimes are ISO 8601 in UTC. Lists are cursor-paged with `?cursor=&limit=`.\n",
    "title": "ais0c analyst API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/qa": {
      "get": {
        "description": "The QA queue across all cases, oldest first: the longest waiting item is first.",
        "operationId": "get_qa_queue_api_v1_qa_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "$ref": "#/components/schemas/QAStatus"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "reason",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "$ref": "#/components/schemas/QAReason"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Reason"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_QAItemSummary_"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "An RFC 9457 problem. `title` is a machine-readable code (`catalog.rule_not_found`); the UI builds the message from it."
          }
        },
        "summary": "Get Qa Queue",
        "tags": [
          "qa"
        ]
      }
    },
    "/api/v1/qa/{item_id}/resolve": {
      "post": {
        "description": "Resolve a QA item and record the operator's feedback for its case, in one transaction.\n\nAn item that is already resolved is a 409 (`qa.already_resolved`) and neither the item nor\nthe feedback changes.",
        "operationId": "resolve_qa_item_route_api_v1_qa__item_id__resolve_post",
        "parameters": [
          {
            "in": "path",
            "name": "item_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Item Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QAResolveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QAResolveAnswer"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "An RFC 9457 problem. `title` is a machine-readable code (`catalog.rule_not_found`); the UI builds the message from it."
          }
        },
        "summary": "Resolve Qa Item Route",
        "tags": [
          "qa"
        ]
      }
    }
  }
}